Play with implementing a DSL for the Robot World commands
The `command` macro allows chaining simple helper functions to define top-level
block-world commands. Each 'step' is passed all the arguments of the command.
Therefore, we must play on the arguments a little to get the desired behavior.
An example of this is the clear-src-block and clear-dest-block; we know which
argument is going to be the src and dest blocks, and we call the clear-block
helper accordingly.

This makes the calling syntax incredibly compact and expressive:
Choose clear helper function names and understand your arguments and you will
get easy-to-understand commands